move common functionality into a common function
