+++ angular-client/src/app/add-contact/add-contact.component.ts
@@ -5,6 +5,7 @@ import { Router } from "@angular/router";
 
 import { Contact } from "../../shared/models/contact";
 import { ContactService } from "../../shared/services/contact.service";
+import { convertToISODateTime } from "../../shared/utils";
 
 /**
  * AddContactComponent is the form that adds a new contact.
@@ -31,32 +32,16 @@ export class AddContactComponent {
 	});
 
 	/**
-	 * Converts a date to an ISO format.
-	 *
-	 * @param dateString The date being converted.
-	 * @returns The converted date.
-	 */
-	convertToISODateTime(dateString: string | null | undefined): string | null | undefined {
-		if (dateString === null || dateString === undefined) {
-			return dateString;
-		}
-		const date = new Date(dateString);
-		date.setUTCHours(0, 0, 0, 0); // Set time to 00:00:00.000 in UTC
-		const isoString = date.toISOString(); // Get the ISO string, e.g., "2024-11-04T00:00:00.000Z"
-		return isoString.replace("Z", "-00:00"); // Replace the Z with -00:00
-	}
-
-	/**
 	 * Handles submission of the form.
 	 */
 	handleSubmit() {
 		const contact: Contact = {
-			Birthday: this.convertToISODateTime(this.contactForm.value.Birthday),
+			Birthday: convertToISODateTime(this.contactForm.value.Birthday),
 			CreatedAt: null,
 			Email: this.contactForm.value.Email,
 			FirstName: this.contactForm.value.FirstName!,
 			ID: null,
-			LastContacted: this.convertToISODateTime(this.contactForm.value.LastContacted),
+			LastContacted: convertToISODateTime(this.contactForm.value.LastContacted),
 			LastName: this.contactForm.value.LastName!,
 			PhoneNumber: this.contactForm.value.PhoneNumber,
 		};
